requirements.txt
# Core AI/ML Libraries
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
ultralytics>=8.0.0
mediapipe>=0.10.5

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
flask>=2.3.0
flask-admin>=1.6.0
jinja2>=3.1.0

# Database
pymongo>=4.4.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.1.0
firebase-admin>=6.2.0

# IoT and Communication
paho-mqtt>=1.6.0
requests>=2.31.0
websockets>=11.0.0

# Image Processing and ML (Optional - for advanced features)
# tensorflow>=2.13.0
# keras>=2.13.0
# onnx>=1.14.0
# onnxruntime>=1.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development and Testing (Optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.0.0

# Cloud Services (Optional)
# boto3>=1.28.0
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.17.0

# Security
cryptography>=41.0.0
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.0

# Configuration
pyyaml>=6.0.0

# Logging and Monitoring (Optional)
# loguru>=0.7.0
